# openapi.yaml
openapi: 3.0.3
info:
  title: API NetWish
  description: API for recommendation films and series, based on the user's preferences in TMDB.
  contact:
    email: dev@netWish.com
  version: 2.0.0
servers:
  - url: http://localhost:3000
    description: Local server
    variables:
      protocol:
        default: http
        enum:
          - http
          - https
      server:
        default: localhost
...
paths:
  /users:
    get:
      security:
        - bearerAuth: []
      tags:
        - User
      summary: Get all users
      description: Get all users
      operationId: getUsers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
          $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Forbidden
        '404':
          description: Not found
        '500':
          description: Internal server error
    post:
      tags:
        - User
      summary: Create a user
      description: Create a user
      operationId: createUser
      requestBody:
        description: User to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            example:
              firstName: John
              lastName: Doe
              email: 'john.doe@outlook.com'
              password: '123456'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                firstName: John
                lastName: Doe
                email: 'john.doe@outlook.com'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '500':
          description: Internal server error
  /users/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - User
      summary: Get a user
      description: Get a user
      operationId: getUser
      parameters:
        - name: id
          in: path
          description: User id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                firstName: John
                lastName: Doe
                email: 'john.doe@outlook.com'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
          $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Forbidden
        '404':
          description: Not found
        '500':
          description: Internal server error
    put:
      security:
        - bearerAuth: []
      tags:
        - User
      summary: Update a user
      description: Update a user
      operationId: updateUser
      parameters:
        - name: id
          in: path
          description: User id
          required: true
          schema:
            type: string
      requestBody:
        description: User to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            example:
                firstName: John
                lastName: Doe
                email: 'john.doe@outlook.com'
                password: '123456'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                firstName: John
                lastName: Doe
                email: 'john.doe@outlook.com'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
          $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Forbidden
        '404':
          description: Not found
        '500':
          description: Internal server error
    delete:
      security:
        - bearerAuth: []
      tags:
        - User
      summary: Delete a user
      description: Delete a user
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          description: User id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              example:
                response: User deleted
        '400':
          description: Bad request
        '401':
          description: Unauthorized
          $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Forbidden
        '404':
          description: Not found
        '500':
          description: Internal server error
  /users/login:
    post:
      tags:
        - User
      summary: Login a user
      description: Login a user
      operationId: loginUser
      requestBody:
        description: User to login
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            example:
              email: 'john.doe@outlook.com'
              password: '123456'
      responses:
        '200':
          description: OK
          content:
            application/json:
              example:
                response: User logged in
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          description: The auto-generated id of the user
        firstName:
          type: string
          description: The first name of the user
        lastName:
          type: string
          description: The last name of the user
        email:
          type: string
          description: The email of the user
          format: email
        password:
          type: string
          description: The password of the user
        admin:
          type: boolean
          description: The admin of the user
          default: false
...
    UnauthorizedError:
      description: Access token is missing or invalid